refactor: configuration 파일 분리 통합
